+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2026-01-04
+
+### Added
+- Confirmation dialog when deleting relationships - asks if inverse relationship should also be deleted
+- Automatic Gravatar sideloading when adding email address to person without an image
+
+### Changed
+- Relationship deletion now gives user control over inverse relationship deletion
+- Email addresses added to people without images will automatically check for and load Gravatar
+
 ## [1.4.9] - 2026-01-04
 
 ### Fixed